Extract CMTime+TimeInterval
