+++ AudioBar/AudioBarViewController.swift
@@ -179,15 +179,3 @@ public final class AudioBarViewController: UIViewController, StoreDelegate {
     }
 
 }
-
-extension CMTime {
-
-    init(timeInterval: TimeInterval) {
-        self = CMTime(seconds: timeInterval, preferredTimescale: 44100)
-    }
-
-    var timeInterval: TimeInterval {
-        return CMTimeGetSeconds(self)
-    }
-
-}
