+++ AudioBar/CMTime+TimeInterval.swift
@@ -1,0 +1,13 @@
+import CoreMedia
+
+extension CMTime {
+
+    init(timeInterval: TimeInterval) {
+        self = CMTime(seconds: timeInterval, preferredTimescale: 44100)
+    }
+
+    var timeInterval: TimeInterval {
+        return CMTimeGetSeconds(self)
+    }
+
+}
